requirements.txt
amqp==2.6.1
asgiref==3.7.2
billiard==3.5.0.5
cffi==1.15.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==4.2.5
django-filter==23.3
djangorestframework==3.14.0
dnspython==2.4.2
gevent==23.9.1
greenlet==3.0.0rc3
kombu==4.6.11
prompt-toolkit==3.0.39
pycparser==2.21
python-dateutil==2.8.2
pytz==2023.3.post1
redis==5.0.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
vine==1.3.0
wcwidth==0.2.6
zope.event==5.0
zope.interface==6.0
